Remove wrapper functions and simplify agent state management
- Remove execute_action, monitor_command_completion, and start_status_monitoring wrapper functions from agent.rs
- Add rule-specific execute_rule_action functions directly in rule modules (when.rs and diff_timeout.rs)
- Update agents.rs to call agent.start_status_monitoring() directly
- Remove exports from mod.rs
- Final simplification completing the agent state management refactoring
