+++ src/agent/mod.rs
@@ -3,5 +3,5 @@ mod agent;
 pub mod agents;
 
 // Re-export for convenience
-pub use agent::{Agent, AgentStatus, execute_action, execute_entry, start_status_monitoring};
+pub use agent::{Agent, AgentStatus};
 pub use agents::Agents;
