+++ src/trigger/mod.rs
@@ -1,3 +1,4 @@
+pub mod execution;
 pub mod periodic;
 pub mod startup;
 pub mod triggers;
