+++ src/trigger/periodic.rs
@@ -5,7 +5,6 @@ use tokio::task::JoinHandle;
 use tokio::time::interval;
 
 use crate::agent::Agents;
-use crate::agent::execute_entry;
 use crate::config::trigger::{Trigger, TriggerType};
 
 /// Periodic task manager responsible for handling periodic entries
@@ -44,7 +43,7 @@ impl Periodic {
                 entry_clone.name
             );
             let agent = agents_clone.get_agent_by_index(0);
-            if let Err(e) = execute_entry(&entry_clone, &agent).await {
+            if let Err(e) = entry_clone.execute(&agent).await {
                 eprintln!(
                     "❌ Error executing startup periodic entry '{}': {}",
                     entry_clone.name, e
@@ -62,7 +61,7 @@ impl Periodic {
                     Ok(true) => {
                         // Execute on available agent
                         let agent = agents_clone.get_agent_by_index(0);
-                        if let Err(e) = execute_entry(&entry_clone, &agent).await {
+                        if let Err(e) = entry_clone.execute(&agent).await {
                             eprintln!(
                                 "❌ Error executing periodic entry '{}': {}",
                                 entry_clone.name, e
